requirements.txt
psutil>=5.8.0
python-valve>=0.2.1
requests>=2.25.1
beautifulsoup4>=4.9.3